requirements.txt
Flask==2.0.3
pyjwt==2.3.0
